Added width resizing, more code for attaching/detaching
Haven't booted with this even, so probably quite a lot of errors but figure i'd commit anyway. Close button for floating panes are also in now. 

Plus adjusting the bottom pane in a sidebar will snap it to the top properly again. Will need to rework that function tomorrow to allow changing order and snapping others but the last one too.