Update a lot of code for the new panes, optimize and clean
The existing panes function well UI wise, but the actual functionality is still broken due to changed IDCs or so. Still cleaning up there and getting rid of weird code. 

Dragging should be more direct (perFrame instead of uiSleep)

Closing windows works again and selecting a window that's already open will bring it to the top.