Fix dragging using a delayed starting position
This will make it so any drag action that uses mouse position, will use the position when the button was pressed, instead of the position after the delay. In practice this means that you don't need to keep your mouse still while pressing a button to start proper dragging making selection/dragging feel A LOT better.